requirements.txt
Django==4.1.10
Pillow==9.4.0
azure-storage-blob==12.13.0
boto3==1.26.24
django-storages==1.12.3
djangorestframework==3.14.0
mysqlclient==2.1.1
python-dotenv==0.19.2
requests==2.31.0
whitenoise==6.0.0
